# ci/src/main/java/eu/wohlben/qits/ci/dto/CiRunDto.java
package eu.wohlben.qits.ci.dto;

import eu.wohlben.qits.ci.entity.CiRunStatus;
import eu.wohlben.qits.ci.entity.CiTriggerType;
import java.time.Instant;
import java.util.List;

/**
 * A CI run as returned to clients — the recorded green/red for one (push, branch). {@code steps} is
 * populated only on the single-run endpoint (with output), null in run listings.
 *
 * <p>{@code live} is the step currently executing and what it has printed so far, read from the
 * in-memory relay rather than from the database: non-null only on the single-run endpoint, and only
 * while {@code status} is {@code RUNNING}. Steps are persisted at their end, so mid-run {@code
 * steps} legitimately holds fewer entries than the pipeline declared, and {@code live} is what says
 * which step the gap belongs to instead of leaving it looking like a run with missing steps.
 *
 * <p><b>{@code QUEUED} is a run that has been accepted and not started</b>, and a client must treat
 * it as non-terminal: it has no {@code finishedAt}, no {@code daemonVersion}, no steps and no {@code
 * live} — there is nothing running to be live about — and it will move on its own, so keep polling.
 * The two active statuses are {@code QUEUED} and {@code RUNNING}; the three terminal ones are {@code
 * SUCCESS}, {@code FAILED} and {@code CONFIG_ERROR}, and a cancelled run is {@code FAILED}.
 *
 * <p>{@code daemonVersion} is the {@code qits-ci-daemon} build every one of this run's containers
 * ran, pinned once at run creation — so the row records forever what produced its results.
 *
 * <p>The four <b>provenance</b> fields say what caused the run. {@code triggerType} is {@code
 * POST_RECEIVE} or {@code EVENT}; {@code configPath} is the committed file that declared the
 * pipeline, which on an event-triggered run identifies <em>which</em> {@code
 * .config/qits/ci-event-*.yml} matched; {@code triggerEventId} and {@code triggerEventName} are the
 * event that caused it, null on every push. They are exposed here because the run API is where an
 * operator reads a run's provenance from outside — no client renders them yet, and that is a later,
 * small follow-up rather than a gap.
 */
public record CiRunDto(
    String id,
    String repoId,
    String branch,
    String commitSha,
    CiRunStatus status,
    Instant createdAt,
    Instant finishedAt,
    String daemonVersion,
    CiTriggerType triggerType,
    String triggerEventId,
    String triggerEventName,
    String configPath,
    List<CiStepDto> steps,
    CiLiveStepDto live) {}